translator/requirements.txt
langdetect
python-dotenv
openai